- Add UIManager, GameManager and merge Player.cs - Get player through GameManager using GetPlayerOne method

+++ Assets/Scripts/General/IDamageable.cs
@@ -3,5 +3,5 @@
     // Called by Player (for AIEnemy) and by AIEnemy (for Building)
     bool IsDead();
     // Called by Player (for AIEnemy) and by AIEnemy (for Building)
-    void TakeDamage(int damage, AttackType attacktype);
+    void TakeDamage(float damage, AttackType attacktype);
 }
